fix(agent-status): address Claude review on TokenCache (PR #115 round 1)
Two findings from Claude Code Review on PR #115, both fixed:

- F1 [MEDIUM] formatTokens imported from sibling component instead of
  shared util (TokenCache.tsx:9). Promoted formatTokens out of
  BudgetMetrics.tsx into a new src/features/agent-status/utils/format.ts;
  updated BudgetMetrics, BudgetMetrics.test, and TokenCache to import
  from there. ContextBucket has its own M-aware formatTokens — left
  alone as out-of-scope per SCOPE BOUNDARY.

- F2 [LOW] StatCell label/value/hint trio rendered as <span>s with no
  semantic association for screen readers (TokenCache.tsx:82-105).
  Promoted the outer grid to <dl> and rewrote each cell as <dt> + <dd>
  + <dd>. Tailwind classes preserved → zero visual change. Satisfies
  WCAG 1.3.1.

Local codex verify on the staged diff: pass (cycle 1, confidence 0.93).
Connector posted +1 reaction on PR #115 — no inline findings.

Pattern KB updates (atomic with the code fix):
- accessibility.md +finding 12; ref_count 1 → 2; last_updated 2026-04-30.
- module-boundaries.md created (code-quality, finding 1).
- docs/reviews/CLAUDE.md index updated for both.

Processed-Claude-Ids: 4351391170
Closes-Codex-Threads: (none)
Pattern-Files-Touched: accessibility.md
Pattern-Files-Touched: module-boundaries.md
Pattern-Files-Touched: docs/reviews/CLAUDE.md
